+++ CHANGELOG.md
@@ -18,6 +18,11 @@ The API consists of all public Java types from `com.atlassian.performance.tools.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/master%0Drelease-3.5.0
 
+### Added
+- Add `AppImpactTest.dataset`. Fix [JPERF-510].
+
+[JPERF-510]: https://ecosystem.atlassian.net/browse/JPERF-510
+
 ## [3.5.0] - 2019-12-09
 [3.5.0]: https://bitbucket.org/atlassian/jira-performance-tests/branches/compare/release-3.5.0%0Drelease-3.4.1